+++ src/PhpWord/Reader/Word2007/AbstractPart.php
@@ -151,16 +151,29 @@ abstract class AbstractPart
         } elseif ($headingDepth !== null) {
             // Heading or Title
             $textContent = null;
-            $nodes = $xmlReader->getElements('w:r', $domNode);
+            $nodes = $xmlReader->getElements('w:r|w:ins/w:r|w:del/w:r', $domNode);
             if ($nodes->length === 1) {
-                $textContent = htmlspecialchars($xmlReader->getValue('w:t', $nodes->item(0)), ENT_QUOTES, 'UTF-8');
+                if($xmlReader->elementExists('w:delText', $nodes->item(0))) {
+                    $textContent = htmlspecialchars($xmlReader->getValue('w:delText', $nodes->item(0)), ENT_QUOTES, 'UTF-8');
+                } else {
+                    $textContent = htmlspecialchars($xmlReader->getValue('w:t', $nodes->item(0)), ENT_QUOTES, 'UTF-8');
+                }
             } else {
                 $textContent = new TextRun($paragraphStyle);
                 foreach ($nodes as $node) {
                     $this->readRun($xmlReader, $node, $textContent, $docPart, $paragraphStyle);
                 }
             }
-            $parent->addTitle($textContent, $headingDepth);
+            $element = $parent->addTitle($textContent, $headingDepth);
+            //add track changes to title
+            if ($xmlReader->elementExists('w:ins|w:del', $domNode)) {
+                $nodesInsDel  = $xmlReader->getElements('w:ins|w:del', $domNode);
+                $nodeChanges = $nodesInsDel->item(0);
+                $type = ($nodeChanges->nodeName == 'w:del') ? TrackChange::DELETED : TrackChange::INSERTED;
+                $author = $nodeChanges->getAttribute('w:author');
+                $date = \DateTime::createFromFormat('Y-m-d\TH:i:s\Z', $nodeChanges->getAttribute('w:date'));
+                $element->setChangeInfo($type, $author, $date);
+            }
         } else {
             // Text and TextRun
             $textRunContainers = $xmlReader->countElements('w:r|w:ins|w:del|w:hyperlink|w:smartTag', $domNode);
