+++ src/PhpWord/Writer/HTML/Element/TrackChangeTrait.php
@@ -1,0 +1,84 @@
+<?php
+/**
+ * This file is part of PHPWord - A pure PHP library for reading and writing
+ * word processing documents.
+ *
+ * PHPWord is free software distributed under the terms of the GNU Lesser
+ * General Public License version 3 as published by the Free Software Foundation.
+ *
+ * For the full copyright and license information, please read the LICENSE
+ * file that was distributed with this source code. For the full list of
+ * contributors, visit https://github.com/PHPOffice/PHPWord/contributors.
+ *
+ * @see         https://github.com/PHPOffice/PHPWord
+ * @copyright   2010-2018 PHPWord contributors
+ * @license     http://www.gnu.org/licenses/lgpl.txt LGPL version 3
+ */
+
+namespace PhpOffice\PhpWord\Writer\HTML\Element;
+
+
+use PhpOffice\PhpWord\Element\TrackChange;
+/**
+ * TrackChange trait for HTML writer
+ *
+ */
+trait TrackChangeTrait
+{
+    /**
+     * writes the track change opening tag
+     *
+     * @return string the HTML, an empty string if no track change information
+     */
+    private function writeTrackChangeOpening()
+    {
+        $changed = $this->element->getTrackChange();
+        if ($changed == null) {
+            return '';
+        }
+
+        $content = '';
+        if (($changed->getChangeType() == TrackChange::INSERTED)) {
+            $content .= '<ins data-phpword-prop=\'';
+        } elseif ($changed->getChangeType() == TrackChange::DELETED) {
+            $content .= '<del data-phpword-prop=\'';
+        }
+
+        $changedProp = array('changed' => array('author'=> $changed->getAuthor(), 'id'    => $this->element->getElementId()));
+        if ($changed->getDate() != null) {
+            $changedProp['changed']['date'] = $changed->getDate()->format('Y-m-d\TH:i:s\Z');
+        }
+        $content .= json_encode($changedProp);
+        $content .= '\' ';
+        $content .= 'title="' . $changed->getAuthor();
+        if ($changed->getDate() != null) {
+            $dateUser = $changed->getDate()->format('Y-m-d H:i:s');
+            $content .= ' - ' . $dateUser;
+        }
+        $content .= '">';
+
+        return $content;
+    }
+
+    /**
+     * writes the track change closing tag
+     *
+     * @return string the HTML, an empty string if no track change information
+     */
+    private function writeTrackChangeClosing()
+    {
+        $changed = $this->element->getTrackChange();
+        if ($changed == null) {
+            return '';
+        }
+
+        $content = '';
+        if (($changed->getChangeType() == TrackChange::INSERTED)) {
+            $content .= '</ins>';
+        } elseif ($changed->getChangeType() == TrackChange::DELETED) {
+            $content .= '</del>';
+        }
+
+        return $content;
+    }
+}
